DocumentReviewer/requirements.txt
python-docx>=0.8.11
pypdf>=3.17.1
pydantic>=2.5.2
python-dotenv>=1.0.0
openai>=1.12.0 
langchain==0.3.20
langchain_anthropic==0.3.9
langchain_community==0.3.19
langchain_core==0.3.41
langchain_groq==0.2.4
langchain_ollama==0.2.3
langchain_openai==0.3.7
python-dotenv==1.0.1
rich==13.9.4
streamlit==1.43.1
markdown==3.5.2 
duckduckgo-search==7.5.1
langgraph==0.3.5